+++ src/black/trans.py
@@ -794,6 +794,8 @@ class StringMerger(StringTransformer, CustomSplitMapMixin):
                 - The set of all string prefixes in the string group is of
                   length greater than one and is not equal to {"", "f"}.
                 - The string group consists of raw strings.
+                - The string group would merge f-strings with different quote types
+                  and internal quotes.
                 - The string group is stringified type annotations. We don't want to
                   process stringified type annotations since pyright doesn't support
                   them spanning multiple string values. (NOTE: mypy, pytype, pyre do
@@ -820,6 +822,8 @@ class StringMerger(StringTransformer, CustomSplitMapMixin):
 
                 i += inc
 
+        QUOTE = line.leaves[string_idx].value[-1]
+
         num_of_inline_string_comments = 0
         set_of_prefixes = set()
         num_of_strings = 0
@@ -841,6 +845,19 @@ class StringMerger(StringTransformer, CustomSplitMapMixin):
                 return TErr("StringMerger does NOT merge raw strings.")
 
             set_of_prefixes.add(prefix)
+
+            if (
+                "f" in prefix
+                and leaf.value[-1] != QUOTE
+                and (
+                    "'" in leaf.value[len(prefix) + 1 : -1]
+                    or '"' in leaf.value[len(prefix) + 1 : -1]
+                )
+            ):
+                return TErr(
+                    "StringMerger does NOT merge f-strings with different quote types"
+                    "and internal quotes."
+                )
 
             if id(leaf) in line.comments:
                 num_of_inline_string_comments += 1